misc/HtmlTestSuite/results/test-0301.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head>
<meta charset="utf-8"/>
<title>Motion Path Module Level 1: calc values</title>
<link href="mailto:ericwilligers@chromium.org" rel="author" title="Eric Willigers"/>
<link href="https://drafts.fxtf.org/motion-1/" rel="help"/>
<link href="https://drafts.csswg.org/css-values-3/#calc-notation" rel="help"/>
<meta content="calc values are supported in offset properties." name="assert"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<style>
  #target {
    font-size: 20px;
  }
</style>
</head>
<body>
<div id="target"/>
<script>
'use strict';

test(function(){
  target.style = 'offset-position: calc(30px + 20%) calc(-200px + 8em + 100%);';
  assert_equals(getComputedStyle(target).offsetPosition, 'calc(20% + 30px) calc(100% - 40px)');
}, 'offset-position supports calc');

test(function(){
  target.style = 'offset-path: ray(calc(1turn - 100grad) closest-side);';
  assert_equals(getComputedStyle(target).offsetPath, 'ray(270deg closest-side)');
}, 'offset-path supports calc');

test(function(){
  target.style = 'offset-distance: calc(-100px + 50%);';
  assert_equals(getComputedStyle(target).offsetDistance, 'calc(50% - 100px)');
}, 'offset-distance supports calc');

test(function(){
  target.style = 'offset-rotate: auto calc(1turn - 100grad);';
  assert_equals(getComputedStyle(target).offsetRotate, 'auto 270deg');
}, 'offset-rotate supports calc');

test(function(){
  target.style = 'offset-anchor: calc(30px + 20%) calc(-200px + 8em + 100%);';
  assert_equals(getComputedStyle(target).offsetAnchor, 'calc(20% + 30px) calc(100% - 40px)');
}, 'offset-anchor supports calc');

</script>
</body>
</html>
